+++ Public/Invoke/Invoke-NinjaOneDeviceScript.ps1
@@ -31,6 +31,10 @@ function Invoke-NinjaOneDeviceScript {
         [Parameter()]
         [string]$runAs
     )
+    if ($Script:NRAPIConnectionInformation.AuthMode -eq 'Client Credentials') {
+        throw ('This function is not available when using client_credentials authentication. Please report this to api@ninjarmm.com.')
+        exit 1
+    }
     try {
         $Resource = "v2/device/$($deviceId)/script/run"
         $RunRequest = @{