+++ Public/Update/Update-NinjaOneOrganisationDocument.ps1
@@ -21,6 +21,10 @@ function Update-NinjaOneOrganisationDocument {
         [Parameter(Mandatory = $true)]
         [object]$organisationDocument
     )
+    if ($Script:NRAPIConnectionInformation.AuthMode -eq 'Client Credentials') {
+        throw ('This function is not available when using client_credentials authentication. Please report this to api@ninjarmm.com.')
+        exit 1
+    }
     try {
         $Resource = "v2/organization/$organisationId/document/$documentId"
         $RequestParams = @{